chore: refactors all react components to seperate files [BREAKING]
- creates components subdir in app
- moves sub components to components subdir
